moved Exceptions to new imports and added dataproxy example
