+++ examples/dataproxy/file.csv
@@ -1,0 +1,5 @@
+id,value
+a,0
+b,1
+c,2
+d,3
