+++ src/foundry_dev_tools/Exceptions/__init__.py
@@ -1,0 +1,4 @@
+from foundry_dev_tools.Exceptions import DataProxyExceptions
+
+
+__all__ = ["DataProxyExceptions"]
